refactor: water to body namespace
